refactor: moving the key vault client
